Break up helper.py into python module
