+++ bot/helper/cli/__init__.py
@@ -1,0 +1,3 @@
+import click
+@click.group()
+def cli(): pass
